requirements.txt
absl-py==2.1.0
accelerate==0.30.1
addict==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
einops==0.8.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
facexlib==0.3.0
filelock==3.13.1
filterpy==1.4.5
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.1.3
future==1.0.0
grpcio==1.64.1
huggingface-hub==0.20.3
icecream==2.1.3
idna==3.6
imageio==2.34.1
imgaug==0.4.0
importlib_metadata==7.1.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
keybert==0.8.4
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.2
llvmlite==0.42.0
lmdb==1.4.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numba==0.59.1
numpy==1.26.4
openai-clip==1.0.1
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==5.27.0
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.17.2
pyiqa==0.1.11
pyparsing==3.1.2
python-dateutil==2.8.2
pytorch-ignite==0.5.0.post2
pytorch-lightning==2.2.5
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
rich==13.7.0
safetensors==0.4.2
scikit-image==0.23.2
scikit-learn==1.4.1.post1
scipy==1.12.0
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
sentence-transformers==2.3.1
sentencepiece==0.1.99
shapely==2.0.4
six==1.16.0
stack-data==0.6.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.3.0
tifffile==2024.5.22
timm==1.0.3
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.0+cu121
torchaudio==2.2.0+cu121
torchmetrics==1.4.0.post0
torchvision==0.17.0+cu121
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.37.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
yapf==0.40.2
yarl==1.9.4
zipp==3.19.2